requirement.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.1.0
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.11.0
dashscope==1.22.2
dataclasses-json==0.6.7
debugpy==1.8.11
decorator==5.1.1
distro==1.9.0
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.10.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2025.3.2
gradio==5.23.0
gradio_client==1.8.0
greenlet==3.1.1
groovy==0.1.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.1
idna==3.10
importlib_metadata==8.5.0
intel-openmp==2023.1.0
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
langchain==0.3.19
langchain-anthropic==0.3.10
langchain-community==0.3.18
langchain-core==0.3.49
langchain-openai==0.3.7
langchain-text-splitters==0.3.6
langsmith==0.2.10
libclang==14.0.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2023.1.0
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.65.2
opencv-python==4.11.0.86
orjson==3.10.13
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pip==25.0.1
platformdirs==4.3.6
ply==3.11
prompt_toolkit==3.0.48
propcache==0.3.0
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydantic-settings==2.8.1
pydub==0.25.1
pygame==2.6.1
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.2.0
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
pywin32==307
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
ruff==0.11.2
safehttpx==0.1.6
safetensors==0.5.3
scipy==1.15.1
semantic-version==2.10.0
setuptools==76.1.0
shellingham==1.5.4
sip==6.7.12
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.46.1
sympy==1.13.3
tbb==2021.13.1
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.21.1
tomli==2.0.1
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.3
typer==0.15.2
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2025.2
unicodedata2==15.1.0
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
yarl==1.18.3
zhipuai==2.1.5.20250106
zipp==3.21.0
